auto refable
(cherry picked from commit c428e7030eeb68096bfcf3e05e5dfda42e30945a)

+++ build.sh
@@ -1,10 +1,17 @@
 
-cd compiler
-rm -rf goldenFile
-mvn package
-cp target/sric-1.0-SNAPSHOT.jar ../bin/sric-1.0.jar
-cd ..
+
+if [ "$1" = "mvn" ]; then
+    cd compiler
+    rm -rf goldenFile
+    mvn package
+    cp target/sric-1.0-SNAPSHOT.jar ../bin/sric-1.0.jar
+    cd ..
+else
+    cd compiler
+    sh build.sh
+    cd ..
+fi
 
 bin/sric ./library/std/module.scm
-
+bin/sric ./library/test/module.scm
